+++ lib/helpers/index.ts
@@ -1,3 +1,3 @@
 import * as binary from './binary';
-import * as multipart from './multipart';
+import * as multipart from './multiplex';
 export { binary, multipart };
